@@ -1,77 +1,62 @@
 import "./App.css";
-import React, { Component } from "react";
+import React, { useState } from "react";
 import Counters from "./component/counters";
 import NavBar from "./component/navBar";
-class App extends Component {
-  // constructor() {
-  //   super();
-  //   console.log("App - Constructor");
-  // }
-  // componentDidMount() {
-  //   console.log("App - componentDidMount");
-  // }
-  // componentDidUpdate(prevProps, prevState) {
-  //   console.log("App - componentDidUpdate");
-  //   console.log("prevProps - ", prevProps);
-  //   console.log("prevState - ", prevState);
-  // }
-  state = {
-    counters: [
-      { id: 1, value: 4 },
-      { id: 2, value: 0 },
-      { id: 3, value: 2 },
-      { id: 4, value: 0 },
-    ],
+
+const App = () => {
+  const [counters, setCounters] = useState([
+    { id: 1, value: 4 },
+    { id: 2, value: 0 },
+    { id: 3, value: 2 },
+    { id: 4, value: 0 },
+  ]);
+
+  const [navColorState, setNavColorState] = useState(
+    "navbar navbar-light bg-light"
+  );
+
+  const handleDelete = (counterId) => {
+    const newCounters = counters.filter((c) => c.id !== counterId);
+    setCounters(newCounters);
   };
-  handleDelete = (counterId) => {
-    const newCounter = this.state.counters.filter(
-      (counter) => counter.id !== counterId
+
+  const handleIncrement = (counter) => {
+    const newCounters = counters.map((c) =>
+      c.id === counter.id ? { ...c, value: c.value + 1 } : c
     );
-    this.setState({ counters: newCounter });
+    setCounters(newCounters);
   };
 
-  handleIncrement = (counter) => {
-    const index = this.state.counters.indexOf(counter);
-    const counters = [...this.state.counters];
-    counters[index] = { ...counter };
-    counters[index].value++;
-    this.setState({ counters });
-  };
-
-  handleDecrement = (counter) => {
-    console.log(counter);
-    const index = this.state.counters.indexOf(counter);
-    const counters = [...this.state.counters];
-    counters[index] = { ...counter };
-    counters[index].value--;
-    this.setState({ counters });
-  };
-
-  handleReset = () => {
-    const counters = this.state.counters.map((c) => {
-      c.value = 0;
-      return c;
-    });
-    this.setState({ counters });
-  };
-
-  render() {
-    // console.log("App - Render");
-    return (
-      <React.Fragment>
-        <NavBar
-          counterCount={this.state.counters.filter((c) => c.value > 0).length}
-        />
-        <Counters
-          onReset={this.handleReset}
-          onDelete={this.handleDelete}
-          onIncrement={this.handleIncrement}
-          onDecrement={this.handleDecrement}
-          counters={this.state.counters}
-        />
-      </React.Fragment>
+  const handleDecrement = (counter) => {
+    const newCounters = counters.map((c) =>
+      c.id === counter.id ? { ...c, value: c.value - 1 } : c
     );
-  }
-}
+    setCounters(newCounters);
+  };
+
+  const handleReset = () => {
+    const resetCounters = counters.map((c) => ({ ...c, value: 0 }));
+    setCounters(resetCounters);
+  };
+
+  const onChangeColor = () => {
+    const newColor =
+      navColorState === "navbar navbar-light bg-dark"
+        ? "navbar navbar-light bg-light"
+        : "navbar navbar-light bg-dark";
+    setNavColorState(newColor);
+  };
+
+  return (
+    <>
+      <NavBar
+        counterCount={counters.filter((c) => c.value > 0).length}
+        navClass={navColorState}
+      />
+      <Counters
+      />
+    </>
+  );
+};
 
 export default App;
