@@ -1,55 +1,52 @@
 import React, { Component } from "react";
+import { increment, decrement, removeCounter } from "../features/counterSlice";
+import { useDispatch, useSelector } from "react-redux";
 
-class Counter extends Component {
-  //   constructor() {
-  //     super();
-  //     this.handleIncrement = this.handleIncrement.bind(this);
-  //   }
+const Counter = (props) => {
+  const {id,counter} = props
+    const dispatch = useDispatch();
 
-  render() {
-    // console.log("counter - rendered");
     return (
       <div className="row">
         {/* <div className="col-1">
           <span className="m-2">counter# {this.props.counter.id}</span>
         </div> */}
         <div className="col-1">
-          <span className={this.getBadgeClasses()}>{this.formatCount()}</span>
+          <span className={getBadgeClasses(counter.value)}>{formatCount(counter.value)}</span>
         </div>
         <div className="col">
           <button
-            onClick={() => this.props.onIncrement(this.props.counter)}
+            onClick={()=> {dispatch(increment(counter.id))}}
             className="btn btn-secondary m-2"
           >
             +
           </button>
           <button
-            onClick={() => this.props.onDecrement(this.props.counter)}
+            onClick={()=> {dispatch(decrement(counter.id))}}
             className="btn btn-secondary m-2"
-            disabled={this.props.counter.value > 0 ? false : true}
+            disabled={counter.value > 0 ? false : true}
           >
             -
           </button>
           <button
-            onClick={() => this.props.onDelete(this.props.counter.id)}
+            onClick={() => dispatch(removeCounter(counter.id))}
             className="btn btn-danger btn-sm m-2"
           >
             X
           </button>
-        </div>
+          </div>
       </div>
     );
-  }
-  getBadgeClasses() {
+}
+  function getBadgeClasses(count) {
     let classes = "badge m-2 badge-";
-    classes += this.props.counter.value === 0 ? "warning" : "primary";
+    classes += count === 0 ? "warning" : "primary";
     return classes;
-  }
+  };
 
-  formatCount() {
-    const { value: count } = this.props.counter;
+  function formatCount(count) {
     return count === 0 ? "Zero" : count;
   }
-}
+
 
 export default Counter;
