@@ -1,33 +1,39 @@
-import React, { Component } from "react";
+import React from "react";
 import Counter from "./counter";
-class Counters extends Component {
-//   componentDidUpdate(prevProps, prevState) {
-//     console.log("App - componentDidUpdate");
-//     console.log("prevProps - ", prevProps);
-//     console.log("prevState - ", prevState);
-//   }
-  render() {
-    const { onReset, onDelete, onIncrement, counters, onDecrement } =
-      this.props;
-    // console.log("counters - rendered");
+import {useDispatch, useSelector} from 'react-redux'
+import { toggleColor } from "../features/changeColorSlice";
+import { addCounter } from "../features/counterSlice";
+const Counters = (
+) => {
+  const counters = useSelector(state => state.counterReducer);
+  const dispatch = useDispatch();
+  return (
+    <div>
+      <button onClick={()=>{dispatch(addCounter())}} className="btn btn-primary btn-sm m-2">
+        Add Counter
+      </button>
 
-    return (
-      <div>
-        <button onClick={onReset} className="btn btn-primary btn-sm m-2">
-          Reset
-        </button>
-        {counters.map((counter) => (
-          <Counter
-            key={counter.id}
-            onDelete={onDelete}
-            onIncrement={onIncrement}
-            onDecrement={onDecrement}
-            counter={counter}
-          />
-        ))}
-      </div>
-    );
-  }
-}
+      {counters.map((counter) => (
+        <Counter
+          key={counter.id}
+          counter={counter}
+        />
+      ))}
+
+      {/* <button onClick={onChangeColor} className="btn btn-primary btn-sm m-2">
+        Change Navbar Color
+      </button> */}
+
+      <button
+        onClick={() => {
+          dispatch(toggleColor())
+        }}
+        className="btn btn-primary btn-sm m-2"
+      >
+        Change Color
+      </button>
+    </div>
+  );
+};
 
 export default Counters;
